# platform/openapi.yaml
openapi: 3.1.0
info:
  title: clickbus-iris
  version: 1.0.0
  description: Clickbus Iris platform API — authentication, organization management, microservice tokens, and metrics ingestion.
  license:
    name: Apache-2.0
servers:
  - url: https://iris.clickbus.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie set after successful sign-in.
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Internal server error
    ErrorResponseWithError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Unauthorized
...
    SignupRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane@example.com
        password:
          type: string
          format: password
          minLength: 8
          description: Must contain at least one uppercase letter, one lowercase letter, one number, and one special character.
          example: "P@ssw0rd!"
    SignupResponse:
      type: object
      required:
        - message
        - user
      properties:
        message:
          type: string
          example: User created successfully. Please check your email to verify your account.
        user:
          type: object
          required:
            - id
            - email
            - name
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            name:
              type: string
...
    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          format: uuid
    Verify2FARequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 6
          description: 6-digit verification code.
          example: "123456"
...
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    RefreshedTokenResponse:
      type: object
      required:
        - accessToken
        - expiresIn
        - tokenType
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          example: 3600
        tokenType:
          type: string
          example: Bearer
    RevokeTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: Required unless revokeAll is true.
        revokeAll:
          type: boolean
          description: When true, revoke all refresh tokens for the authenticated user.
...
  responses:
    BadRequest:
      description: Bad request - validation error or missing required fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or session invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /api/auth/signup:
    post:
      operationId: signup
      summary: Register a new user
      description: |
        Creates a new user account with email/password credentials and sends a verification email.
        Rate limit tier: authRateLimit.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Forgot Password
  # ──────────────────────────────────────────────
...
  /api/auth/verify-email:
    post:
      operationId: verifyEmail
      summary: Verify email address
      description: |
        Confirms a user's email using the verification token sent during signup.
        Rate limit tier: strictAuthRateLimit.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        "200":
          description: Email verified successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Email verified successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Verify 2FA (session-based)
  # ──────────────────────────────────────────────
  /api/auth/verify-2fa:
    post:
      operationId: verify2FA
      summary: Verify a 2FA code (authenticated)
      description: |
        Verifies a 6-digit 2FA code for the currently authenticated user session.
        Rate limit tier: authRateLimit.
      tags:
        - auth
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Verify2FARequest"
      responses:
        "200":
          description: 2FA verification successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: 2FA verification successful
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Complete 2FA (login flow)
  # ──────────────────────────────────────────────
...
  /api/microservices/refresh:
    post:
      operationId: refreshMicroserviceToken
      summary: Refresh a microservice access token
      description: |
        Exchanges a valid refresh token for a new JWT access token (1 hour).
        The user must still have an active session and the refresh token must match the session user.
        Rate limit tier: apiRateLimit.
      tags:
        - microservices
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: New access token issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshedTokenResponse"
        "400":
          description: Missing refresh token or no organization context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithError"
        "401":
          description: Invalid refresh token or session mismatch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithError"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  # ──────────────────────────────────────────────
  # Microservices - Revoke Token
  # ──────────────────────────────────────────────
  /api/microservices/revoke:
    post:
      operationId: revokeMicroserviceToken
      summary: Revoke refresh token(s)
      description: |
        Revokes a specific refresh token or all refresh tokens for the authenticated user (when `revokeAll` is true).
        Rate limit tier: apiRateLimit.
      tags:
        - microservices
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeTokenRequest"
      responses:
        "200":
          description: Token(s) revoked.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Refresh token revoked successfully
        "400":
          description: Missing refresh token (when revokeAll is not set).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithError"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithError"

  # ──────────────────────────────────────────────
  # Users - Avatar
  # ──────────────────────────────────────────────
...